fix(saves,engine): restore-prepare clobber + flatten silent-drop + time-sync i64 overflow (round-3 audit)
- save_archive restore_prepare: a restore zip carrying BOTH `<name>` and
  `sdimg_<name>` would rename the bare file over the existing sdimg_
  image (std::fs::rename overwrites on Unix / errors on Windows),
  destroying the real image. Add the same exists-guard backup_finalize
  uses for the inverse rename.
- save_archive flatten_wrapper_subdirs: used filter_map(|e| e.ok()) then
  remove_dir_all the source — a per-entry read error silently dropped a
  file and then deleted it (incomplete backup, no error). Propagate the
  error instead.
- engine ps5_time_sync_route: (new_unix - target) could overflow i64 for
  an extreme request-controlled target (debug panic / release wrong
  boolean). Compute the drift in i128.

Found in the round-3 audit (save/restore + engine-handler reviews). The
reviews also verified zip-slip, symlink, savedata-folder preservation,
partial-write ordering, job-registry races, and path-traversal handling
all correct.

+++ engine/crates/ps5upload-engine/src/main.rs
@@ -1592,7 +1592,13 @@ async fn ps5_time_sync_route(
             // away from what we asked for. Either the SDK stub is a
             // no-op on this firmware, or the underlying syscall is
             // refusing silently. Surface so the UI can warn.
-            let stub_no_op = v.ok && v.new_unix >= 0 && (v.new_unix - target).abs() > 5;
+            // i128 subtraction: `target` is request-controlled (i64 from
+            // JSON) and `v.new_unix` is payload-derived, so `new_unix -
+            // target` in i64 can overflow (e.g. target = i64::MIN) —
+            // debug panics, release wraps to a wrong boolean. Widen so the
+            // drift comparison is always correct.
+            let stub_no_op =
+                v.ok && v.new_unix >= 0 && (v.new_unix as i128 - target as i128).abs() > 5;
             let resp = TimeSyncResp {
                 ok: v.ok,
                 err_code: v.err_code,
